Move CurrencyCard onChange to useExchange
